+++ lib/src/collection_extensions.dart
@@ -198,9 +198,22 @@ extension CollectionExtension<T> on Iterable<T> {
   /// Returns a new lazy [Iterable] with elements that are created by
   /// calling `toElement` on each element of this `Iterable` in
   /// iteration order.
-  Iterable<K> mapWithIndex<K>(K Function(int index, T item) toElement) sync* {
+  Iterable<K> mapWithIndex<K>(
+    K Function(
+      int index,
+      T item,
+      bool isFirst,
+      bool isLast,
+    )
+        toElement,
+  ) sync* {
     for (var i = 0; i < length; i++) {
-      yield toElement(i, elementAt(i));
+      yield toElement(
+        i,
+        elementAt(i),
+        i == 0,
+        i == length - 1,
+      );
     }
   }
 
@@ -215,6 +228,13 @@ extension CollectionExtension<T> on Iterable<T> {
       } else {
         yield elementAt(i);
       }
+    }
+  }
+
+  Iterable<Iterable<T>> splitOnPagesBy(int count) sync* {
+    final totalPages = (length / count).floor() + (length % count > 0 ? 1 : 0);
+    for (var i = 0; i < totalPages; i++) {
+      yield skip(i * count).take(count).cast<T>();
     }
   }
 }